Extract incidents skeleton
