+++ Client/src/Pages/Incidents/index.jsx
@@ -1,43 +1,12 @@
 import { useState, useEffect } from "react";
 import { useSelector } from "react-redux";
-import {
-  ButtonGroup,
-  Stack,
-  Skeleton,
-  Typography,
-  Button,
-} from "@mui/material";
+import { ButtonGroup, Stack, Typography, Button } from "@mui/material";
 import { networkService } from "../../main";
 import { useTheme } from "@emotion/react";
 import Select from "../../Components/Inputs/Select";
 import IncidentTable from "./IncidentTable";
 import "./index.css";
-
-/**
- * Renders a skeleton layout.
- *
- * @returns {JSX.Element}
- */
-const SkeletonLayout = () => {
-  const theme = useTheme();
-
-  return (
-    <>
-      <Stack direction="row" alignItems="center" gap={theme.spacing(6)}>
-        <Skeleton variant="rounded" width={150} height={34} />
-        <Skeleton variant="rounded" width="15%" height={34} />
-        <Skeleton
-          variant="rounded"
-          width="20%"
-          height={34}
-          sx={{ ml: "auto" }}
-        />
-      </Stack>
-      <Skeleton variant="rounded" width="100%" height={300} />
-      <Skeleton variant="rounded" width="100%" height={100} />
-    </>
-  );
-};
+import SkeletonLayout from "./skeleton";
 
 const Incidents = () => {
   const theme = useTheme();
